% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generator.R
\name{appendDeadlinesToJobData}
\alias{appendDeadlinesToJobData}
\title{Generiere Deadline-Instanz von 2-Maschinen-Flow-Shop.}
\usage{
appendDeadlinesToJobData(jobData, bufferType, maxBufferSize)
}
\description{
Generiere Deadline-Instanz von 2-Maschinen-Flow-Shop.
}
