% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToDueDateInstance.R
\name{convertToDueDateInstance}
\alias{convertToDueDateInstance}
\title{Generiere Deadline-Instanz von 2-Maschinen-Flow-Shop
aus normaler Job-Data-Datei}
\usage{
convertToDueDateInstance(fileName)
}
\description{
Generiere Deadline-Instanz von 2-Maschinen-Flow-Shop
aus normaler Job-Data-Datei
}
