% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generator.R
\name{generateNormalProblemInstance}
\alias{generateNormalProblemInstance}
\title{Generiere normalverteilte Instanz von 2-Maschinen-Flow-Shop.}
\usage{
generateNormalProblemInstance(numberOfJobs, m, s)
}
\arguments{
\item{numberOfJobs}{Anzahl der Jobs.}

\item{aLimit}{Zweielementiger Vektor.}

\item{bLimit}{Zweielementiger Vektor (die Grenzen). Wenn nur eine Zahl, dann konstant diese.}

\item{bufferLimit}{Zweielementiger Vektor (Grenzen). Wenn
1 (Skalar), dann konstant 1. Wenn negativer Skalar, dann wie bei M1.}
}
\description{
Generiere normalverteilte Instanz von 2-Maschinen-Flow-Shop.
}
